jacoco/src/test/java/org/sonar/plugins/jacoco/JaCoCoPluginTest.java
/*
 * Sonar JaCoCo Plugin
 * Copyright (C) 2010 SonarSource
 * dev@sonar.codehaus.org
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with this program; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
 */

package org.sonar.plugins.jacoco;

import static org.hamcrest.Matchers.greaterThan;
import static org.hamcrest.Matchers.is;
import static org.hamcrest.Matchers.notNullValue;
import static org.junit.Assert.assertThat;

import org.junit.Before;
import org.junit.Test;

/**
 * @author Evgeny Mandrikov
 */
public class JaCoCoPluginTest {
  private JaCoCoPlugin plugin = new JaCoCoPlugin();

  @Before
  public void setUp() {
    plugin = new JaCoCoPlugin();
  }

  @Test
  public void testPluginDefition() {
    assertThat(plugin.getKey(), is("jacoco"));
    assertThat(plugin.getName(), notNullValue());
    assertThat(plugin.getDescription(), notNullValue());
    assertThat(plugin.toString(), is("jacoco"));
  }

  @Test
  public void testExtensions() {
    assertThat(plugin.getExtensions().size(), greaterThan(0));
  }
}
